docs: add simplified deploy of sidecar using default config
This change does the following:
- Make it so the `job` label comes directly from the config. This makes
  the behaviour consistent with `PodMonitoring` in the GKE operator and
  alse is covered in go/run-gmp-scrape
- Add a cloudbuild script for a simple deploy that doesn't involve
  secret manager
- Update docs to clarify that the secret manager steps are only
  necessary if you want to customize the config.
- Refactored the readme so split out the default config and custom
  config journeys
- Create new Cloud Run service config for when secret manager is not
  needed

+++ clean-up-cloud-run.sh
@@ -18,7 +18,11 @@ SA_NAME="run-gmp-sa"
 REGION="us-east1"
 
 gcloud run services delete run-gmp-sidecar-service --region ${REGION} --quiet
-gcloud secrets delete run-gmp-config
+# Delete secret if we created it before
+if gcloud secrets list --filter="name ~ .*run-gmp-config.*" | grep run-gmp-sidecar
+then
+  gcloud secrets delete run-gmp-config
+fi
 gcloud artifacts repositories delete run-gmp \
   --location=${REGION} \
   --quiet
